Move test classes into namespace
